lab-3/zad-1/requirements.txt
pip~=20.0.2
wheel~=0.34.2
setuptools~=46.1.3
pybuilder~=0.12.5
six~=1.14.0
progressbar2